Fix component clutter
Каждый файл с несколькими компонентами разбит на отдельные файлы
