+++ src/components/Footer/Footer.module.sass
@@ -34,24 +34,6 @@ div.socialMedia
   gap: 25px
 
 
-div.facebook
-  div.iconBox
-    justify-content: center
-
-
-div.socialMediaContainer
-  display: inline-flex
-  align-items: center
-  gap: 10px
-
-  a.link
-    color: $black
-    text-decoration: none
-    font-size: 20px
-
-    +underline($black, 1px)
-
-
 @media screen and (max-width: 1800px)
   footer
     padding-inline: 53px
@@ -80,16 +62,3 @@ div.socialMediaContainer
 
     p.description
       font-size: 16px
-
-  div.socialMediaContainer
-    gap: 5px
-
-    a.link
-      font-size: 16px
-
-  div.iconBox
-    width: 20px
-    height: 20px
-
-    img.icon
-      scale: .8
