+++ src/components/Footer/component/SocialMediaLink.modules.sass
@@ -1,0 +1,31 @@
+div.facebook
+  div.iconBox
+    justify-content: center
+
+
+div.socialMediaContainer
+  display: inline-flex
+  align-items: center
+  gap: 10px
+
+  a.link
+    color: $black
+    text-decoration: none
+    font-size: 20px
+
+    +underline($black, 1px)
+
+
+@media screen and (max-width: 450px)
+  div.socialMediaContainer
+    gap: 5px
+
+    a.link
+      font-size: 16px
+
+    div.iconBox
+      width: 20px
+      height: 20px
+
+      img.icon
+        scale: .8
