+++ src/components/Footer/component/index.js
@@ -1,0 +1,1 @@
+export {default as SocialMediaLink} from './SocialMediaLink'
