+++ src/pages/Home/components/BuildYourTour/components/FormInputFields/component/index.js
@@ -1,0 +1,1 @@
+export {default as InputTemplate} from './InputTemplate'
