+++ src/pages/Home/components/Header/Header.module.sass
@@ -65,11 +65,11 @@ header.header
       background: $orange100
 
 @media screen and (max-width: 1024px)
-  header>p.description
+  header.header > p.description
       width: 45%
 
 @media screen and (max-width: 700px)
-  header
+  header.header
     height: 488px
     gap: 40px
