+++ src/pages/Home/components/Reviews/Reviews.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import SectionTitle from "../SectionTitle/SectionTitle";
 import {UsersFeedbackData} from "../../../../data/UsersFeedbackData";
+import {ReviewTemplate} from "./components";
 import style from './Reviews.module.sass'
 
 function Reviews() {
@@ -14,31 +15,6 @@ function Reviews() {
             </div>
         </section>
     );
-}
-
-function ReviewTemplate({responseInfo, index}) {
-    return(
-        <div className={`${style.response} ${style.response}__${index}`}>
-            <p className={style.text}>
-                {responseInfo.text}
-            </p>
-
-            <div className={style.userInfo}>
-                <div className={style.description}>
-                    <h2 className={style.name}>
-                        {responseInfo.userInfo.name}
-                    </h2>
-
-                    <p className={style.tourName}>
-                        {`Тур: ${responseInfo.userInfo.tour}`}
-                    </p>
-                </div>
-
-                <img className={style.profilePicture}
-                     src={responseInfo.userInfo.profilePicture} alt={responseInfo.userInfo.name}/>
-            </div>
-        </div>
-    )
 }
 
 export default Reviews;
