+++ src/pages/Home/components/Reviews/components/index.js
@@ -1,0 +1,1 @@
+export {default as ReviewTemplate} from './ReviewTemplate'
