+++ src/pages/Home/components/TourPhotos/TourPhotos.js
@@ -2,6 +2,7 @@ import React from 'react';
 import SectionTitle from "../SectionTitle/SectionTitle";
 import {TourGalleryData} from "../../../../data/TourGalleryData";
 import style from './TourPhotos.module.sass';
+import {PhotosRowTemplate} from "./components";
 
 function TourPhotos() {
     return (
@@ -10,32 +11,22 @@ function TourPhotos() {
 
             <div className={style.photosContainer}>
                 <PhotosRowTemplate rowInfo={{
-                    row: style.firstRow,
+                    row: 'firstRow',
                     pictures: TourGalleryData.slice(0,4)
                 }}/>
 
                 <PhotosRowTemplate rowInfo={{
-                    row: style.secondRow,
+                    row: 'secondRow',
                     pictures: TourGalleryData.slice(4,9)
                 }}/>
 
                 <PhotosRowTemplate rowInfo={{
-                    row: style.thirdRow,
+                    row: 'thirdRow',
                     pictures: TourGalleryData.slice(9,13)
                 }}/>
             </div>
         </section>
     );
-}
-
-function PhotosRowTemplate({rowInfo}) {
-    return(
-        <div className={`${style.photosRow} ${rowInfo.row}`}>
-            {rowInfo.pictures.map(picture =>
-                <img src={picture.src} alt={picture.name} key={picture.name}
-                    className={style.image} />)}
-        </div>
-    )
 }
 
 export default TourPhotos;
